// src/login/dto/user.dto.ts
export class UserDto {
    uid: string;
    displayName: string;
    fullName: string;
    avatar: string;
    cover: string;
    email: string;
    status: string;
    XpPoints: number;
    // friends: friendsRequest[];
    wins: number;
    loses: number;
    draws: number;
    globaRank: number;
    twoFactorAuthenticationSecret: string;
    isTwoFactorAuthenticationEnabled: boolean;
  
  }
 